requirements-gpu.txt
cython==0.28.3
h5py==2.8.0
imgaug==0.2.6
keras==2.2.0
numpy==1.14.5
Pillow==5.2.0
scikit-image==0.14.0
scipy==1.1.0
tensorflow-gpu==1.8.0
grpcio-tools
googleapis-common-protos